+++ R/feature_selection.R
@@ -35,6 +35,9 @@ run_feature_selection <- function(input_data,
   }
 
   require_vip("feature selection")
+  for (package in c("Boruta", "corrr", "ranger")) {
+    require_optional_package(package, "feature selection")
+  }
 
   # check if outlier cleaning has been applied
   if ("Target_Original" %in% colnames(input_data)) {
@@ -307,6 +310,7 @@ target_corr_fn <- function(data,
 vip_rf_fn <- function(data,
                       seed = 123) {
   require_vip("random forest feature selection")
+  require_optional_package("ranger", "random forest feature selection")
 
   rf_mod <- parsnip::rand_forest(mode = "regression", trees = 100) %>%
     parsnip::set_engine("ranger", importance = "impurity")
@@ -416,8 +420,16 @@ vip_cubist_fn <- function(data,
 boruta_fn <- function(data,
                       iterations = 100,
                       seed = 123) {
+  require_optional_package("Boruta", "Boruta feature selection")
+  require_optional_package("ranger", "Boruta feature selection")
+
   set.seed(seed)
-  Boruta::Boruta(Target ~ ., data = data, maxRuns = iterations) %>%
+  Boruta::Boruta(
+    Target ~ .,
+    data = data,
+    maxRuns = iterations,
+    getImp = Boruta::getImpRfZ
+  ) %>%
     Boruta::getSelectedAttributes()
 }